requirements.txt
h5py==3.12.1
ipython==8.31.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
scikit-image==0.25.0
scipy==1.15.1
seaborn==0.13.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
tqdm==4.67.1
